Move ChangesetEvents & Changeset state computation to enterprise
We should always err on the side of having more code in enterprise than
in open.

This moves a lot of the actual code (not type definitions) out of
`internal/campaigns/types.go` to multiple files in
`enterprise/internal/campaigns`.

This is kinda related to #9101, because CalcCounts lives in enterprise
and sharing code between open and enterprise can get tricky.
