+++ internal/campaigns/types_test.go
@@ -1,11 +1,8 @@
 package campaigns
 
 import (
-	"sort"
 	"testing"
 	"time"
-
-	"github.com/google/go-cmp/cmp/cmpopts"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketserver"
@@ -258,666 +255,4 @@ func TestChangesetEvents(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestChangesetEventsState(t *testing.T) {
-	tests := []struct {
-		sortedEvents ChangesetEvents
-		want         ChangesetState
-	}{
-		{
-			sortedEvents: ChangesetEvents{},
-			want:         ChangesetStateOpen,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindGitHubClosed},
-			},
-			want: ChangesetStateClosed,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindBitbucketServerDeclined},
-			},
-			want: ChangesetStateClosed,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindGitHubClosed},
-				{Kind: ChangesetEventKindGitHubReopened},
-			},
-			want: ChangesetStateOpen,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindBitbucketServerDeclined},
-				{Kind: ChangesetEventKindBitbucketServerReopened},
-			},
-			want: ChangesetStateOpen,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindGitHubClosed},
-				{Kind: ChangesetEventKindGitHubReopened},
-				{Kind: ChangesetEventKindGitHubClosed},
-			},
-			want: ChangesetStateClosed,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindBitbucketServerDeclined},
-				{Kind: ChangesetEventKindBitbucketServerReopened},
-				{Kind: ChangesetEventKindBitbucketServerDeclined},
-			},
-			want: ChangesetStateClosed,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindGitHubMerged},
-			},
-			want: ChangesetStateMerged,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindBitbucketServerMerged},
-			},
-			want: ChangesetStateMerged,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindGitHubMerged},
-				// Merged is a final state. Events after should be ignored.
-				{Kind: ChangesetEventKindGitHubClosed},
-			},
-			want: ChangesetStateMerged,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				{Kind: ChangesetEventKindBitbucketServerMerged},
-				// Merged is a final state. Events after should be ignored.
-				{Kind: ChangesetEventKindBitbucketServerDeclined},
-			},
-			want: ChangesetStateMerged,
-		},
-		{
-			sortedEvents: ChangesetEvents{
-				// GitHub emits Closed and Merged events at the same time.
-				// We want to report Merged.
-				{Kind: ChangesetEventKindGitHubClosed},
-				{Kind: ChangesetEventKindGitHubMerged},
-			},
-			want: ChangesetStateMerged,
-		},
-	}
-
-	for i, tc := range tests {
-		if have, want := tc.sortedEvents.State(), tc.want; have != want {
-			t.Errorf("%d: wrong state. have=%s, want=%s", i, have, want)
-		}
-	}
-}
-
-func TestChangesetEventsReviewState(t *testing.T) {
-	now := time.Now().UTC().Truncate(time.Microsecond)
-	daysAgo := func(days int) time.Time { return now.AddDate(0, 0, -days) }
-	ghReview := func(t time.Time, login, state string) *ChangesetEvent {
-		return &ChangesetEvent{
-			Kind: ChangesetEventKindGitHubReviewed,
-			Metadata: &github.PullRequestReview{
-				UpdatedAt: t,
-				State:     state,
-				Author: github.Actor{
-					Login: login,
-				},
-			},
-		}
-	}
-
-	ghReviewDismissed := func(t time.Time, login, reviewer string) *ChangesetEvent {
-		return &ChangesetEvent{
-			Kind: ChangesetEventKindGitHubReviewDismissed,
-			Metadata: &github.ReviewDismissedEvent{
-				CreatedAt: t,
-				Actor:     github.Actor{Login: login},
-				Review: github.PullRequestReview{
-					Author: github.Actor{
-						Login: reviewer,
-					},
-				},
-			},
-		}
-	}
-
-	bbsActivity := func(t time.Time, login string, kind ChangesetEventKind) *ChangesetEvent {
-		return &ChangesetEvent{
-			Kind: kind,
-			Metadata: &bitbucketserver.Activity{
-				CreatedDate: timeToUnixMilli(t),
-				User: bitbucketserver.User{
-					Name: login,
-				},
-			},
-		}
-	}
-
-	tests := []struct {
-		events ChangesetEvents
-		want   ChangesetReviewState
-	}{
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(0), "user1", "APPROVED"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(1), "user1", "APPROVED"),
-				ghReview(daysAgo(0), "user1", "COMMENTED"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(1), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(0), "user1", "COMMENTED"),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(1), "user1", "APPROVED"),
-				ghReview(daysAgo(0), "user1", "PENDING"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(1), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(0), "user1", "PENDING"),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(2), "user1", "APPROVED"),
-				ghReview(daysAgo(1), "user1", "CHANGES_REQUESTED"),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(2), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(1), "user1", "APPROVED"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(0), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(0), "user2", "APPROVED"),
-				ghReview(daysAgo(0), "user3", "APPROVED"),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(3), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(2), "user2", "APPROVED"),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(3), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(2), "user2", "APPROVED"),
-				ghReview(daysAgo(0), "user1", "APPROVED"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(1), "user1", "CHANGES_REQUESTED"),
-				ghReviewDismissed(daysAgo(0), "user2", "user1"),
-			},
-			want: ChangesetReviewStatePending,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(2), "user1", "CHANGES_REQUESTED"),
-				ghReviewDismissed(daysAgo(1), "user2", "user1"),
-				ghReview(daysAgo(0), "user1", "CHANGES_REQUESTED"),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(2), "user1", "CHANGES_REQUESTED"),
-				ghReviewDismissed(daysAgo(1), "user2", "user1"),
-				ghReview(daysAgo(0), "user3", "APPROVED"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(1), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(0), "user1", "DISMISSED"),
-			},
-			want: ChangesetReviewStatePending,
-		},
-		{
-			events: ChangesetEvents{
-				ghReview(daysAgo(2), "user1", "CHANGES_REQUESTED"),
-				ghReview(daysAgo(1), "user1", "DISMISSED"),
-				ghReview(daysAgo(0), "user3", "APPROVED"),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerApproved),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerReviewed),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerApproved),
-				bbsActivity(daysAgo(1), "user2", ChangesetEventKindBitbucketServerReviewed),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerApproved),
-				bbsActivity(daysAgo(1), "user2", ChangesetEventKindBitbucketServerReviewed),
-				bbsActivity(daysAgo(0), "user3", ChangesetEventKindBitbucketServerApproved),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerApproved),
-				bbsActivity(daysAgo(1), "user2", ChangesetEventKindBitbucketServerReviewed),
-				bbsActivity(daysAgo(0), "user2", ChangesetEventKindBitbucketServerApproved),
-			},
-			want: ChangesetReviewStateApproved,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerApproved),
-				bbsActivity(daysAgo(1), "user1", ChangesetEventKindBitbucketServerUnapproved),
-			},
-			want: ChangesetReviewStatePending,
-		},
-		{
-			events: ChangesetEvents{
-				bbsActivity(daysAgo(2), "user1", ChangesetEventKindBitbucketServerApproved),
-				bbsActivity(daysAgo(1), "user1", ChangesetEventKindBitbucketServerUnapproved),
-				bbsActivity(daysAgo(0), "user1", ChangesetEventKindBitbucketServerReviewed),
-			},
-			want: ChangesetReviewStateChangesRequested,
-		},
-	}
-
-	for i, tc := range tests {
-		sort.Sort(tc.events)
-		have, err := tc.events.reviewState()
-		if err != nil {
-			t.Fatalf("got error: %s", err)
-		}
-
-		if have, want := have, tc.want; have != want {
-			t.Errorf("%d: wrong reviewstate. have=%s, want=%s", i, have, want)
-		}
-	}
-}
-
-func TestComputeGithubCheckState(t *testing.T) {
-	now := time.Now().UTC().Truncate(time.Microsecond)
-	commitEvent := func(minutesSinceSync int, context, state string) *ChangesetEvent {
-		commit := &github.CommitStatus{
-			Context:    context,
-			State:      state,
-			ReceivedAt: now.Add(time.Duration(minutesSinceSync) * time.Minute),
-		}
-		event := &ChangesetEvent{
-			Kind:     ChangesetEventKindCommitStatus,
-			Metadata: commit,
-		}
-		return event
-	}
-	checkRun := func(id, status, conclusion string) github.CheckRun {
-		return github.CheckRun{
-			ID:         id,
-			Status:     status,
-			Conclusion: conclusion,
-		}
-	}
-	checkSuiteEvent := func(minutesSinceSync int, id, status, conclusion string, runs ...github.CheckRun) *ChangesetEvent {
-		suite := &github.CheckSuite{
-			ID:         id,
-			Status:     status,
-			Conclusion: conclusion,
-			ReceivedAt: now.Add(time.Duration(minutesSinceSync) * time.Minute),
-		}
-		suite.CheckRuns.Nodes = runs
-		event := &ChangesetEvent{
-			Kind:     ChangesetEventKindCheckSuite,
-			Metadata: suite,
-		}
-		return event
-	}
-
-	lastSynced := now.Add(-1 * time.Minute)
-	pr := &github.PullRequest{}
-
-	tests := []struct {
-		name   string
-		events []*ChangesetEvent
-		want   ChangesetCheckState
-	}{
-		{
-			name:   "empty slice",
-			events: nil,
-			want:   ChangesetCheckStateUnknown,
-		},
-		{
-			name: "single success",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "SUCCESS"),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-		{
-			name: "success status and suite",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "SUCCESS"),
-				checkSuiteEvent(1, "cs1", "COMPLETED", "SUCCESS", checkRun("cr1", "COMPLETED", "SUCCESS")),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-		{
-			name: "single pending",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "PENDING"),
-			},
-			want: ChangesetCheckStatePending,
-		},
-		{
-			name: "single error",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "ERROR"),
-			},
-			want: ChangesetCheckStateFailed,
-		},
-		{
-			name: "pending + error",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "PENDING"),
-				commitEvent(1, "ctx2", "ERROR"),
-			},
-			want: ChangesetCheckStatePending,
-		},
-		{
-			name: "pending + success",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "PENDING"),
-				commitEvent(1, "ctx2", "SUCCESS"),
-			},
-			want: ChangesetCheckStatePending,
-		},
-		{
-			name: "success + error",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "SUCCESS"),
-				commitEvent(1, "ctx2", "ERROR"),
-			},
-			want: ChangesetCheckStateFailed,
-		},
-		{
-			name: "success x2",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "SUCCESS"),
-				commitEvent(1, "ctx2", "SUCCESS"),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-		{
-			name: "later events have precedence",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "PENDING"),
-				commitEvent(1, "ctx1", "SUCCESS"),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-		{
-			name: "suites with zero runs should be ignored",
-			events: []*ChangesetEvent{
-				commitEvent(1, "ctx1", "SUCCESS"),
-				checkSuiteEvent(1, "cs1", "QUEUED", ""),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			got := computeGitHubCheckState(lastSynced, pr, tc.events)
-			if diff := cmp.Diff(tc.want, got); diff != "" {
-				t.Fatalf(diff)
-			}
-		})
-	}
-}
-
-func TestComputeBitbucketBuildStatus(t *testing.T) {
-	now := time.Now().UTC().Truncate(time.Microsecond)
-	sha := "abcdef"
-	statusEvent := func(minutesSinceSync int, key, state string) *ChangesetEvent {
-		commit := &bitbucketserver.CommitStatus{
-			Commit: sha,
-			Status: bitbucketserver.BuildStatus{
-				State:     state,
-				Key:       key,
-				DateAdded: now.Add(1*time.Second).Unix() * 1000,
-			},
-		}
-		event := &ChangesetEvent{
-			Kind:     ChangesetEventKindBitbucketServerCommitStatus,
-			Metadata: commit,
-		}
-		return event
-	}
-
-	lastSynced := now.Add(-1 * time.Minute)
-	pr := &bitbucketserver.PullRequest{
-		Commits: []*bitbucketserver.Commit{
-			{
-				ID: sha,
-			},
-		},
-	}
-
-	tests := []struct {
-		name   string
-		events []*ChangesetEvent
-		want   ChangesetCheckState
-	}{
-		{
-			name:   "empty slice",
-			events: nil,
-			want:   ChangesetCheckStateUnknown,
-		},
-		{
-			name: "single success",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "SUCCESSFUL"),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-		{
-			name: "single pending",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "INPROGRESS"),
-			},
-			want: ChangesetCheckStatePending,
-		},
-		{
-			name: "single error",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "FAILED"),
-			},
-			want: ChangesetCheckStateFailed,
-		},
-		{
-			name: "pending + error",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "INPROGRESS"),
-				statusEvent(1, "ctx2", "FAILED"),
-			},
-			want: ChangesetCheckStatePending,
-		},
-		{
-			name: "pending + success",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "INPROGRESS"),
-				statusEvent(1, "ctx2", "SUCCESSFUL"),
-			},
-			want: ChangesetCheckStatePending,
-		},
-		{
-			name: "success + error",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "SUCCESSFUL"),
-				statusEvent(1, "ctx2", "FAILED"),
-			},
-			want: ChangesetCheckStateFailed,
-		},
-		{
-			name: "success x2",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "SUCCESSFUL"),
-				statusEvent(1, "ctx2", "SUCCESSFUL"),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-		{
-			name: "later events have precedence",
-			events: []*ChangesetEvent{
-				statusEvent(1, "ctx1", "INPROGRESS"),
-				statusEvent(1, "ctx1", "SUCCESSFUL"),
-			},
-			want: ChangesetCheckStatePassed,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			have := computeBitbucketBuildStatus(lastSynced, pr, tc.events)
-			if diff := cmp.Diff(tc.want, have); diff != "" {
-				t.Fatalf(diff)
-			}
-		})
-	}
-}
-
-func TestChangesetEventsLabels(t *testing.T) {
-	now := time.Now()
-	labelEvent := func(name string, kind ChangesetEventKind, when time.Time) *ChangesetEvent {
-		removed := kind == ChangesetEventKindGitHubUnlabeled
-		return &ChangesetEvent{
-			Kind:      kind,
-			UpdatedAt: when,
-			Metadata: &github.LabelEvent{
-				Actor: github.Actor{},
-				Label: github.Label{
-					Name: name,
-				},
-				CreatedAt: when,
-				Removed:   removed,
-			},
-		}
-	}
-	changeset := func(names []string, updated time.Time) *Changeset {
-		meta := &github.PullRequest{}
-		for _, name := range names {
-			meta.Labels.Nodes = append(meta.Labels.Nodes, github.Label{
-				Name: name,
-			})
-		}
-		return &Changeset{
-			UpdatedAt: updated,
-			Metadata:  meta,
-		}
-	}
-	labels := func(names ...string) []ChangesetLabel {
-		var ls []ChangesetLabel
-		for _, name := range names {
-			ls = append(ls, ChangesetLabel{Name: name})
-		}
-		return ls
-	}
-
-	tests := []struct {
-		name      string
-		changeset *Changeset
-		events    ChangesetEvents
-		want      []ChangesetLabel
-	}{
-		{
-			name: "zero values",
-		},
-		{
-			name:      "no events",
-			changeset: changeset([]string{"label1"}, time.Time{}),
-			events:    ChangesetEvents{},
-			want:      labels("label1"),
-		},
-		{
-			name:      "remove event",
-			changeset: changeset([]string{"label1"}, time.Time{}),
-			events: ChangesetEvents{
-				labelEvent("label1", ChangesetEventKindGitHubUnlabeled, now),
-			},
-			want: []ChangesetLabel{},
-		},
-		{
-			name:      "add event",
-			changeset: changeset([]string{"label1"}, time.Time{}),
-			events: ChangesetEvents{
-				labelEvent("label2", ChangesetEventKindGitHubLabeled, now),
-			},
-			want: labels("label1", "label2"),
-		},
-		{
-			name:      "old add event",
-			changeset: changeset([]string{"label1"}, now.Add(5*time.Minute)),
-			events: ChangesetEvents{
-				labelEvent("label2", ChangesetEventKindGitHubLabeled, now),
-			},
-			want: labels("label1"),
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			have := tc.events.UpdateLabelsSince(tc.changeset)
-			want := tc.want
-			sort.Slice(have, func(i, j int) bool { return have[i].Name < have[j].Name })
-			sort.Slice(want, func(i, j int) bool { return want[i].Name < want[j].Name })
-			if diff := cmp.Diff(have, want, cmpopts.EquateEmpty()); diff != "" {
-				t.Fatal(diff)
-			}
-		})
-	}
-}
-
-func timeToUnixMilli(t time.Time) int {
-	return int(t.UnixNano()) / int(time.Millisecond)
 }
